Moved home and user routes from app.js to individual router files - purely organizational, no change in functionality
